Improve rendering of stats.
We are no longer reading elements from the webpage on each gettimeofday call.
IIUC, reading the element from the rendered webpage froces the browser to render and layout the page each time!
By tracking the number of calls in a javascript variable, there is no need to re-render the page on each call.
In addition, by updating only once per second, we need to rerender the page even less often.
Loading of the level is now much faster.